+ Added read functionality to ngp_linkmasta_flash_device

+++ src/common/types.h
@@ -1,7 +1,9 @@
 #ifndef __TYPES_H__
 #define __TYPES_H__
 
-typedef unsigned long address_t;    //<! Data type representing a storage address
+#include <stdint.h>
+
+typedef uint32_t address_t;    //<! Data type representing a storage address
 
 #endif // __TYPES_H__
